+++ src/classes/metadata/OHIFSeriesMetadata.js
@@ -12,7 +12,7 @@ export class OHIFSeriesMetadata extends SeriesMetadata {
 
     init(study) {
         const series = this.getData();
-
+        // console.log('OHIFSeriesMetadata', series);
         // define "_seriesInstanceUID" protected property...
         Object.defineProperty(this, '_seriesInstanceUID', {
             configurable: false,
@@ -25,5 +25,17 @@ export class OHIFSeriesMetadata extends SeriesMetadata {
         series.instances.forEach((instance) => {
             this.addInstance(new OHIFInstanceMetadata(instance, series, study));
         });
+    }
+
+    isSubSeries() {
+        return this._data.isSubSeries;
+    }
+
+    getSeriesDescription() {
+        return this._data.SeriesDescription;
+    }
+
+    getSeriesNumber() {
+        return this._data.SeriesNumber;
     }
 }