+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Add series split functionality - [@Triet]
+- Add getters function in OHIFSeriesMetadata - [@Triet]
+## Changed
+- Changed addSeries condition in OHIFStudyMetadata - [@Triet]
+
 ## [1.0.8] - 2020-07-17
 ### Fixed
 - Fixed packaged bundle version - [@tunglt1810]